project/requirements.txt
appdirs==1.4.3
Babel==2.4.0
Django==1.11.1
django-phonenumber-field==1.3.0
packaging==16.8
phonenumberslite==8.5.0
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
